pyproject.toml
[project]
name = "agentic-ai"
version = "0.1.0"
description = "Green Agent Math Evaluator for AgentBeats"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agentbeats>=0.1.0",
    "openai>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
]
